+++ tests/e2e/web_test.py
@@ -143,8 +143,8 @@ def test_body_weight_add(driver: webdriver.Chrome) -> None:
 
     page.body_weight_dialog.click_save()
 
-    page.wait_for_table_value(1, date)
-    page.wait_for_table_value(2, weight)
+    page.wait_for_table_value(1, 1, date)
+    page.wait_for_table_value(1, 2, weight)
 
 
 def test_body_weight_edit(driver: webdriver.Chrome) -> None:
@@ -156,22 +156,22 @@ def test_body_weight_edit(driver: webdriver.Chrome) -> None:
     page = BodyWeightPage(driver)
     page.load()
 
-    page.wait_for_table_value(1, date)
-    page.wait_for_table_value(2, weight)
+    page.wait_for_table_value(1, 1, date)
+    page.wait_for_table_value(1, 2, weight)
 
     page.click_edit(0)
     page.body_weight_dialog.set_weight(new_weight)
     page.body_weight_dialog.click_cancel()
 
-    page.wait_for_table_value(1, date)
-    page.wait_for_table_value(2, weight)
+    page.wait_for_table_value(1, 1, date)
+    page.wait_for_table_value(1, 2, weight)
 
     page.click_edit(0)
     page.body_weight_dialog.set_weight(new_weight)
     page.body_weight_dialog.click_save()
 
-    page.wait_for_table_value(1, date)
-    page.wait_for_table_value(2, new_weight)
+    page.wait_for_table_value(1, 1, date)
+    page.wait_for_table_value(1, 2, new_weight)
 
 
 def test_body_weight_delete(driver: webdriver.Chrome) -> None:
@@ -183,19 +183,19 @@ def test_body_weight_delete(driver: webdriver.Chrome) -> None:
     page = BodyWeightPage(driver)
     page.load()
 
-    page.wait_for_table_value(1, date_1)
-    page.wait_for_table_value(2, weight)
+    page.wait_for_table_value(1, 1, date_1)
+    page.wait_for_table_value(1, 2, weight)
 
     page.click_delete(0)
     page.delete_dialog.click_no()
 
-    page.wait_for_table_value(1, date_1)
-    page.wait_for_table_value(2, weight)
+    page.wait_for_table_value(1, 1, date_1)
+    page.wait_for_table_value(1, 2, weight)
 
     page.click_delete(0)
     page.delete_dialog.click_yes()
 
-    page.wait_for_table_value(1, date_2)
+    page.wait_for_table_value(1, 1, date_2)
 
 
 def test_body_fat_add(driver: webdriver.Chrome) -> None:
@@ -227,9 +227,9 @@ def test_body_fat_add(driver: webdriver.Chrome) -> None:
 
     page.body_fat_dialog.click_save()
 
-    page.wait_for_table_value(1, date)
+    page.wait_for_table_value(1, 1, date)
     for i, v in enumerate(values, start=4):
-        page.wait_for_table_value(i, v)
+        page.wait_for_table_value(1, i, v)
 
 
 def test_body_fat_edit(driver: webdriver.Chrome) -> None:
@@ -250,25 +250,25 @@ def test_body_fat_edit(driver: webdriver.Chrome) -> None:
     page = BodyFatPage(driver)
     page.load()
 
-    page.wait_for_table_value(1, date)
+    page.wait_for_table_value(1, 1, date)
     for i, v in enumerate(values, start=4):
-        page.wait_for_table_value(i, v)
+        page.wait_for_table_value(1, i, v)
 
     page.click_edit(0)
     page.body_fat_dialog.set_jp7(new_values)
     page.body_fat_dialog.click_cancel()
 
-    page.wait_for_table_value(1, date)
+    page.wait_for_table_value(1, 1, date)
     for i, v in enumerate(values, start=4):
-        page.wait_for_table_value(i, v)
+        page.wait_for_table_value(1, i, v)
 
     page.click_edit(0)
     page.body_fat_dialog.set_jp7(new_values)
     page.body_fat_dialog.click_save()
 
-    page.wait_for_table_value(1, date)
+    page.wait_for_table_value(1, 1, date)
     for i, v in enumerate(new_values, start=4):
-        page.wait_for_table_value(i, v)
+        page.wait_for_table_value(1, i, v)
 
 
 def test_body_fat_delete(driver: webdriver.Chrome) -> None:
@@ -279,17 +279,17 @@ def test_body_fat_delete(driver: webdriver.Chrome) -> None:
     page = BodyFatPage(driver)
     page.load()
 
-    page.wait_for_table_value(1, date_1)
+    page.wait_for_table_value(1, 1, date_1)
 
     page.click_delete(0)
     page.delete_dialog.click_no()
 
-    page.wait_for_table_value(1, date_1)
+    page.wait_for_table_value(1, 1, date_1)
 
     page.click_delete(0)
     page.delete_dialog.click_yes()
 
-    page.wait_for_table_value(1, date_2)
+    page.wait_for_table_value(1, 1, date_2)
 
 
 def test_menstrual_cycle_add(driver: webdriver.Chrome) -> None:
@@ -319,8 +319,8 @@ def test_menstrual_cycle_add(driver: webdriver.Chrome) -> None:
 
     page.period_dialog.click_save()
 
-    page.wait_for_table_value(1, date)
-    page.wait_for_table_value(2, intensity)
+    page.wait_for_table_value(1, 1, date)
+    page.wait_for_table_value(1, 2, intensity)
 
 
 def test_menstrual_cycle_edit(driver: webdriver.Chrome) -> None:
@@ -335,22 +335,22 @@ def test_menstrual_cycle_edit(driver: webdriver.Chrome) -> None:
     page = MenstrualCyclePage(driver)
     page.load()
 
-    page.wait_for_table_value(1, date)
-    page.wait_for_table_value(2, intensity)
+    page.wait_for_table_value(1, 1, date)
+    page.wait_for_table_value(1, 2, intensity)
 
     page.click_edit(0)
     page.period_dialog.set_period(new_intensity)
     page.period_dialog.click_cancel()
 
-    page.wait_for_table_value(1, date)
-    page.wait_for_table_value(2, intensity)
+    page.wait_for_table_value(1, 1, date)
+    page.wait_for_table_value(1, 2, intensity)
 
     page.click_edit(0)
     page.period_dialog.set_period(new_intensity)
     page.period_dialog.click_save()
 
-    page.wait_for_table_value(1, date)
-    page.wait_for_table_value(2, new_intensity)
+    page.wait_for_table_value(1, 1, date)
+    page.wait_for_table_value(1, 2, new_intensity)
 
 
 def test_menstrual_cycle_delete(driver: webdriver.Chrome) -> None:
@@ -363,19 +363,19 @@ def test_menstrual_cycle_delete(driver: webdriver.Chrome) -> None:
     page = MenstrualCyclePage(driver)
     page.load()
 
-    page.wait_for_table_value(1, date_1)
-    page.wait_for_table_value(2, intensity)
+    page.wait_for_table_value(1, 1, date_1)
+    page.wait_for_table_value(1, 2, intensity)
 
     page.click_delete(0)
     page.delete_dialog.click_no()
 
-    page.wait_for_table_value(1, date_1)
-    page.wait_for_table_value(2, intensity)
+    page.wait_for_table_value(1, 1, date_1)
+    page.wait_for_table_value(1, 2, intensity)
 
     page.click_delete(0)
     page.delete_dialog.click_yes()
 
-    page.wait_for_table_value(1, date_2)
+    page.wait_for_table_value(1, 1, date_2)
 
 
 def test_training_links(driver: webdriver.Chrome) -> None:
@@ -445,19 +445,19 @@ def test_training_delete(driver: webdriver.Chrome) -> None:
     page = TrainingPage(driver)
     page.load()
 
-    page.wait_for_table_value(1, date_1)
-    page.wait_for_table_value(2, routine)
+    page.wait_for_table_value(1, 1, date_1)
+    page.wait_for_table_value(1, 2, routine)
 
     page.click_delete(0)
     page.delete_dialog.click_no()
 
-    page.wait_for_table_value(1, date_1)
-    page.wait_for_table_value(2, routine)
+    page.wait_for_table_value(1, 1, date_1)
+    page.wait_for_table_value(1, 2, routine)
 
     page.click_delete(0)
     page.delete_dialog.click_yes()
 
-    page.wait_for_table_value(1, date_2)
+    page.wait_for_table_value(1, 1, date_2)
 
 
 def test_training_session(driver: webdriver.Chrome) -> None:
@@ -597,7 +597,7 @@ def test_routines_add(driver: webdriver.Chrome) -> None:
 
     page.routines_dialog.click_cancel()
 
-    page.wait_for_table_value(1, name)
+    page.wait_for_table_value(2, 1, name)
 
     page.click_fab()
 
@@ -606,11 +606,11 @@ def test_routines_add(driver: webdriver.Chrome) -> None:
     page.routines_dialog.set_name(new_name)
     page.routines_dialog.click_save()
 
-    page.wait_for_table_value(1, new_name)
+    page.wait_for_table_value(2, 1, new_name)
 
 
 def test_routines_edit(driver: webdriver.Chrome) -> None:
-    name = str(USER.routines[-1].name)
+    name = str(USER.routines[-2].name)
     new_name = "Changed Routine"
 
     assert name != new_name
@@ -619,19 +619,19 @@ def test_routines_edit(driver: webdriver.Chrome) -> None:
     page = RoutinesPage(driver)
     page.load()
 
-    page.wait_for_table_value(1, name)
+    page.wait_for_table_value(1, 1, name)
 
     page.click_edit(0)
     page.routines_dialog.set_name(new_name)
     page.routines_dialog.click_cancel()
 
-    page.wait_for_table_value(1, name)
+    page.wait_for_table_value(1, 1, name)
 
     page.click_edit(0)
     page.routines_dialog.set_name(new_name)
     page.routines_dialog.click_save()
 
-    page.wait_for_table_value(1, new_name)
+    page.wait_for_table_value(1, 1, new_name)
 
 
 def test_routines_delete(driver: webdriver.Chrome) -> None:
@@ -642,17 +642,19 @@ def test_routines_delete(driver: webdriver.Chrome) -> None:
     page = RoutinesPage(driver)
     page.load()
 
-    page.wait_for_table_value(1, name_1)
+    page.wait_for_table_value(1, 1, name_2)
+    page.wait_for_table_value(2, 1, name_1)
 
     page.click_delete(0)
     page.delete_dialog.click_no()
 
-    page.wait_for_table_value(1, name_1)
+    page.wait_for_table_value(1, 1, name_2)
+    page.wait_for_table_value(2, 1, name_1)
 
     page.click_delete(0)
     page.delete_dialog.click_yes()
 
-    page.wait_for_table_value(1, name_2)
+    page.wait_for_table_value(1, 1, name_1)
 
 
 def test_routine_edit_save(driver: webdriver.Chrome) -> None:
@@ -1290,7 +1292,7 @@ def test_exercises_add(driver: webdriver.Chrome) -> None:
 
     page.exercises_dialog.click_cancel()
 
-    page.wait_for_table_value(1, name)
+    page.wait_for_table_value(1, 1, name)
 
     page.click_fab()
 
@@ -1299,7 +1301,7 @@ def test_exercises_add(driver: webdriver.Chrome) -> None:
     page.exercises_dialog.set_name(new_name)
     page.exercises_dialog.click_save()
 
-    page.wait_for_table_value(1, new_name)
+    page.wait_for_table_value(1, 1, new_name)
 
 
 def test_exercises_edit(driver: webdriver.Chrome) -> None:
@@ -1313,19 +1315,19 @@ def test_exercises_edit(driver: webdriver.Chrome) -> None:
     page = ExercisesPage(driver)
     page.load()
 
-    page.wait_for_table_value(1, name)
+    page.wait_for_table_value(1, 1, name)
 
     page.click_edit(0)
     page.exercises_dialog.set_name(new_name)
     page.exercises_dialog.click_cancel()
 
-    page.wait_for_table_value(1, name)
+    page.wait_for_table_value(1, 1, name)
 
     page.click_edit(0)
     page.exercises_dialog.set_name(new_name)
     page.exercises_dialog.click_save()
 
-    page.wait_for_table_value(1, new_name)
+    page.wait_for_table_value(1, 1, new_name)
 
 
 def test_exercises_delete(driver: webdriver.Chrome) -> None:
@@ -1337,17 +1339,17 @@ def test_exercises_delete(driver: webdriver.Chrome) -> None:
     page = ExercisesPage(driver)
     page.load()
 
-    page.wait_for_table_value(1, name_1)
+    page.wait_for_table_value(1, 1, name_1)
 
     page.click_delete(0)
     page.delete_dialog.click_no()
 
-    page.wait_for_table_value(1, name_1)
+    page.wait_for_table_value(1, 1, name_1)
 
     page.click_delete(0)
     page.delete_dialog.click_yes()
 
-    page.wait_for_table_value(1, name_2)
+    page.wait_for_table_value(1, 1, name_2)
 
 
 def test_exercise_delete_workout(driver: webdriver.Chrome) -> None:
@@ -1360,14 +1362,14 @@ def test_exercise_delete_workout(driver: webdriver.Chrome) -> None:
     page = ExercisePage(driver, exercise.id)
     page.load()
 
-    page.wait_for_table_value(1, workout_1)
+    page.wait_for_table_value(1, 1, workout_1)
 
     page.click_delete(0)
     page.delete_dialog.click_no()
 
-    page.wait_for_table_value(1, workout_1)
+    page.wait_for_table_value(1, 1, workout_1)
 
     page.click_delete(0)
     page.delete_dialog.click_yes()
 
-    page.wait_for_table_value(1, workout_2)
+    page.wait_for_table_value(1, 1, workout_2)