+++ CHANGELOG.md
@@ -13,6 +13,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Option to defer exercise in training session
 - Option to add set in training session
 
+### Changed
+
+- List routines sorted by last use
+
 ## [0.3.0] - 2024-01-02
 
 ### Added